Move RGO shuffle helpers into their own files.
